fix1 refactory functions for creating dumb data for testing
