+++ tests/test_milkcocoa_api.py
@@ -6,14 +6,17 @@ import unittest
 from milkcocoa.milkcocoa import Milkcocoa
 from milkcocoa.keys_file_api import MilkCocoaKeys
 
+from dumb_data_creation_scripts import delete_dumb_file, create_dumb_file
+
 
 class TestMilkcocoaApi(unittest.TestCase):
 
     def setUp(self):
+        create_dumb_file()
         self.keys = MilkCocoaKeys()
 
     def tearDown(self):
-        pass
+        delete_dumb_file()
 
     def test_milkcocoa_instance_connection_no_ssl(self):
         credentials = self.keys.get_credentials()
